requirements.txt
evdev==1.7.1
pynput==1.7.7
python-xlib==0.33
rti.connext==7.3.0
six==1.17.0
